# services/core/src/swagger/swagger.yaml
openapi: "3.0.3"
info:
  title: GlobalTransact API
  version: "2.0.0"
  description: |
    International money transfer platform API.

    All endpoints require Bearer JWT authentication unless marked as public.

    **Roles**: client | agent | admin | compliance_officer | super_admin
servers:
  - url: /api/v1
    description: Core API v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        role: { type: string, enum: [client, agent, admin, compliance_officer, super_admin] }
        status: { type: string, enum: [active, suspended, pending] }
        country: { type: string }
        msisdn: { type: string }
        createdAt: { type: string, format: date-time }
...
    Beneficiary:
      type: object
      properties:
        id: { type: string, format: uuid }
        userId: { type: string }
        nickname: { type: string }
        fullName: { type: string }
        msisdn: { type: string }
        country: { type: string }
        currency: { type: string }
        payoutMethod: { type: string, enum: [momo, bank, cash] }
        isFavorite: { type: boolean }
...
security:
  - BearerAuth: []
paths:
  /auth/register:
    post:
      tags: [Authentication]
      security: []
      summary: Register a new client account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, name, password]
              properties:
                email: { type: string, format: email }
                name: { type: string }
                password: { type: string, minLength: 10 }
                country: { type: string }
                msisdn: { type: string }
      responses:
        201: { description: Registered successfully }
        409: { description: Email already registered }
  /auth/login:
    post:
      tags: [Authentication]
      security: []
      summary: Login and receive JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        401: { description: Invalid credentials }
        423: { description: Account locked }
...
  /auth/forgot-password:
    post:
      tags: [Authentication]
      security: []
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
      responses:
        200: { description: Reset email sent (always 200 to prevent enumeration) }
...
  /transfers:
    post:
      tags: [Transfers]
      summary: Initiate a new money transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [recipientName, recipientMsisdn, recipientCountry, sendCurrency, recvCurrency, sendAmount]
              properties:
                recipientName: { type: string }
                recipientMsisdn: { type: string }
                recipientCountry: { type: string }
                sendCurrency: { type: string, enum: [RWF, USD, EUR, GBP, KES, UGX, GHS] }
                recvCurrency: { type: string, enum: [RWF, USD, EUR, GBP, KES, UGX, GHS] }
                sendAmount: { type: number, minimum: 1 }
                fundingMethod: { type: string, enum: [momo, cash, wallet, bank], default: momo }
                payoutMethod: { type: string, enum: [momo, cash, bank], default: momo }
                beneficiaryId: { type: string, format: uuid }
      responses:
        201: { description: Transfer initiated }
        422: { description: Blocked by fraud engine or insufficient funds }
        503: { description: FX rates stale or no liquidity }

    get:
      tags: [Transfers]
      summary: List own transfers
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        200: { description: Transfer list }
...
  /beneficiaries:
    get:
      tags: [Beneficiaries]
      summary: List saved beneficiaries
      responses:
        200: { description: Beneficiaries list }
    post:
      tags: [Beneficiaries]
      summary: Add a new beneficiary
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Beneficiary' }
      responses:
        201: { description: Beneficiary created }
  /beneficiaries/{id}:
    get:
      tags: [Beneficiaries]
      summary: Get beneficiary by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200: { description: Beneficiary }
    put:
      tags: [Beneficiaries]
      summary: Update beneficiary
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200: { description: Updated }
    delete:
      tags: [Beneficiaries]
      summary: Remove beneficiary
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        200: { description: Removed }
...
  /fraud/alerts:
    get:
      tags: [Fraud]
      summary: List fraud alerts (admin/compliance)
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: severity
          schema: { type: string }
      responses:
        200: { description: Alerts list }
  /fraud/alerts/{id}/resolve:
    put:
      tags: [Fraud]
      summary: Resolve a fraud alert
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [status, resolution]
              properties:
                status: { type: string, enum: [RESOLVED, FALSE_POSITIVE, INVESTIGATING] }
                resolution: { type: string }
      responses:
        200: { description: Alert resolved }
...
  /fraud/aml-checks/{userId}:
    post:
      tags: [Fraud]
      summary: Trigger manual AML screen for a user
      parameters:
        - in: path
          name: userId
          required: true
          schema: { type: string }
      responses:
        201: { description: AML check created }
...